+++ src/main/java/jakarta/transaction/HeuristicMixedException.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package javax.transaction;
+package jakarta.transaction;
 
 /**
  * This exception is thrown to report that a heuristic decision was made and
