+++ src/main/java/jakarta/transaction/NotSupportedException.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package javax.transaction;
+package jakarta.transaction;
 
 /**
  * NotSupportedException exception indicates that the request cannot be
