+++ src/main/java/jakarta/transaction/Transaction.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package javax.transaction;
+package jakarta.transaction;
 
 import javax.transaction.xa.XAResource;
 import java.lang.IllegalStateException;
