+++ src/main/java/jakarta/transaction/TransactionManager.java
@@ -14,17 +14,16 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package javax.transaction;
+package jakarta.transaction;
 
-import java.lang.IllegalArgumentException;
 import java.lang.IllegalStateException;
 import java.lang.SecurityException;
 
 /**
- * The UserTransaction interface defines the methods that allow an
- * application to explicitly manage transaction boundaries.
+ * The TransactionManager interface defines the methods that allow an
+ * application server to manage transaction boundaries.
  */
-public interface UserTransaction {
+public interface TransactionManager {
 
     /**
      * Create a new transaction and associate it with the current thread.
@@ -37,7 +36,7 @@ public interface UserTransaction {
      *    encounters an unexpected error condition.
      *
      */
-    void begin() throws NotSupportedException, SystemException;
+    public void begin() throws NotSupportedException, SystemException;
 
     /**
      * Complete the transaction associated with the current thread. When this
@@ -62,14 +61,64 @@ public interface UserTransaction {
      *
      * @exception SystemException Thrown if the transaction manager
      *    encounters an unexpected error condition.
-    */
-    void commit() throws RollbackException,
+     *
+     */
+    public void commit() throws RollbackException,
 	HeuristicMixedException, HeuristicRollbackException, SecurityException,
 	IllegalStateException, SystemException;
 
     /**
+     * Obtain the status of the transaction associated with the current thread.
+     *
+     * @return The transaction status. If no transaction is associated with
+     *    the current thread, this method returns the Status.NoTransaction
+     *    value.
+     *
+     * @exception SystemException Thrown if the transaction manager
+     *    encounters an unexpected error condition.
+     *
+     */
+    public int getStatus() throws SystemException;
+
+    /**
+     * Get the transaction object that represents the transaction
+     * context of the calling thread.
+     *
+     * @return the <code>Transaction</code> object representing the
+     *	  transaction associated with the calling thread.
+     *
+     * @exception SystemException Thrown if the transaction manager
+     *    encounters an unexpected error condition.
+     *
+     */
+    public Transaction getTransaction() throws SystemException;
+
+    /**
+     * Resume the transaction context association of the calling thread
+     * with the transaction represented by the supplied Transaction object.
+     * When this method returns, the calling thread is associated with the
+     * transaction context specified.
+     *
+     * @param tobj The <code>Transaction</code> object that represents the
+     *    transaction to be resumed.
+     *
+     * @exception InvalidTransactionException Thrown if the parameter
+     *    transaction object contains an invalid transaction.
+     *
+     * @exception IllegalStateException Thrown if the thread is already
+     *    associated with another transaction.
+     *
+     * @exception SystemException Thrown if the transaction manager
+     *    encounters an unexpected error condition.
+     */
+    public void resume(Transaction tobj)
+            throws InvalidTransactionException, IllegalStateException,
+            SystemException;
+
+    /**
      * Roll back the transaction associated with the current thread. When this
-     * method completes, the thread is no longer associated with a transaction.
+     * method completes, the thread is no longer associated with a
+     * transaction.
      *
      * @exception SecurityException Thrown to indicate that the thread is
      *    not allowed to roll back the transaction.
@@ -81,8 +130,8 @@ public interface UserTransaction {
      *    encounters an unexpected error condition.
      *
      */
-    void rollback() throws IllegalStateException, SecurityException,
-        SystemException;
+    public void rollback() throws IllegalStateException, SecurityException,
+                            SystemException;
 
     /**
      * Modify the transaction associated with the current thread such that
@@ -96,20 +145,7 @@ public interface UserTransaction {
      *    encounters an unexpected error condition.
      *
      */
-    void setRollbackOnly() throws IllegalStateException, SystemException;
-
-    /**
-     * Obtain the status of the transaction associated with the current thread.
-     *
-     * @return The transaction status. If no transaction is associated with
-     *    the current thread, this method returns the Status.NoTransaction
-     *    value.
-     *
-     * @exception SystemException Thrown if the transaction manager
-     *    encounters an unexpected error condition.
-     *
-     */
-    int getStatus() throws SystemException;
+    public void setRollbackOnly() throws IllegalStateException, SystemException;
 
     /**
      * Modify the timeout value that is associated with transactions started
@@ -126,5 +162,21 @@ public interface UserTransaction {
      *    encounters an unexpected error condition.
      *
      */
-    void setTransactionTimeout(int seconds) throws SystemException;
+    public void setTransactionTimeout(int seconds) throws SystemException;
+
+    /**
+     * Suspend the transaction currently associated with the calling
+     * thread and return a Transaction object that represents the
+     * transaction context being suspended. If the calling thread is
+     * not associated with a transaction, the method returns a null
+     * object reference. When this method returns, the calling thread
+     * is not associated with a transaction.
+     *
+     * @return Transaction object representing the suspended transaction.
+     *
+     * @exception SystemException Thrown if the transaction manager
+     *    encounters an unexpected error condition.
+     *
+     */
+    public Transaction suspend() throws SystemException;
 }
