+++ src/main/java/jakarta/transaction/TransactionSynchronizationRegistry.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package javax.transaction;
+package jakarta.transaction;
 
 /**
  * This interface is intended for use by system level application server
